+++ src/index.test.ts
@@ -40,7 +40,7 @@ describe("responsive()", () => {
   });
 
   it("should add styles in order", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       default: {
         display: "block",
       },
@@ -59,8 +59,53 @@ describe("responsive()", () => {
     );
   });
 
+  it("should work with nested objects", () => {
+    const responsiveDisplay = responsiveStyles.variants({
+      default: {
+        display: "block",
+      },
+      flex: {
+        display: "flex",
+      },
+      inlineBlock: {
+        display: "inline-block",
+      },
+    });
+
+    expect(
+      responsiveDisplay.css({ phone: { inlineBlock: true, flex: true } })
+    ).toBe(
+      `display:block;@media ${mediaQueries.phone}{display:inline-block;display:flex;}`
+    );
+  });
+
+  it("should work insert into dom with nested objects", () => {
+    const responsiveDisplay = responsiveStyles.variants({
+      default: {
+        display: "block",
+      },
+      flex: {
+        display: "flex",
+      },
+      inlineBlock: {
+        display: "inline-block",
+      },
+    });
+
+    expect(
+      responsiveDisplay({ phone: { inlineBlock: true, flex: true } })
+    ).not.toBe("");
+    expect(document.querySelectorAll(`style[data-dash]`).length).toBe(3);
+    expect(document.querySelectorAll(`style[data-dash]`)[1]).toMatchSnapshot(
+      "block"
+    );
+    expect(document.querySelectorAll(`style[data-dash]`)[2]).toMatchSnapshot(
+      "inline-block+flex"
+    );
+  });
+
   it("should provide tokens", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       default: ({ color }) => ({ color: color.white }),
       backgroundColor: ({ color }) => ({ backgroundColor: color.white }),
     });
@@ -70,14 +115,14 @@ describe("responsive()", () => {
   });
 
   it("should return add just the default", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       default: ({ color }) => ({ color: color.white }),
     });
     expect(responsiveDisplay.css()).toBe("color:var(--color-white);");
   });
 
   it("should return empty string for no variant match", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       backgroundColor: ({ color }) => ({ backgroundColor: color.white }),
     });
     expect(responsiveDisplay.css()).toBe("");
@@ -85,7 +130,7 @@ describe("responsive()", () => {
   });
 
   it("should work with style map", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       default: {
         display: "block",
       },
@@ -101,7 +146,7 @@ describe("responsive()", () => {
   });
 
   it("should work without default in style map", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       flex: {
         display: "flex",
       },
@@ -114,7 +159,7 @@ describe("responsive()", () => {
   });
 
   it("should add media queries to style map", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       flex: {
         display: "flex",
       },
@@ -131,7 +176,7 @@ describe("responsive()", () => {
   });
 
   it("should return empty string for misses", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       flex: {
         display: "flex",
       },
@@ -143,8 +188,21 @@ describe("responsive()", () => {
     expect(responsiveDisplay()).toBe("");
   });
 
+  it("should hit base styles", () => {
+    const responsiveDisplay = responsiveStyles.variants({
+      flex: {
+        display: "flex",
+      },
+      inlineBlock: {
+        display: "inline-block",
+      },
+    });
+
+    expect(responsiveDisplay("flex")).not.toBe("");
+  });
+
   it("should insert class into the dom", () => {
-    const responsiveDisplay = responsiveStyles({
+    const responsiveDisplay = responsiveStyles.variants({
       flex: {
         display: "flex",
       },